requirements.txt
# async-generator==.env.test.10
# attrs==22..env.test.0
# backports.functools-lru-cache==.env.test.6.4
# cachetools==5.2.0
# certifi==2022.9.24
# charset-normalizer==2..env.test..env.test
# exceptiongroup==.env.test.0.4
# execnet==.env.test.9.0
Faker==15.3.3
# flake8==5.0.4
# flake8-print==5.0.0
# future==0.18.2
# google-api-core==2.11.0
# google-auth==2.15.0
# grpcio==.env.test.48..env.test
# h11==0.14.0
# importlib-resources==5.10.0
# iniconfig==.env.test..env.test..env.test
# isort==5.10..env.test
jsonschema==4.17.1
# mccabe==0.7.0
# outcome==.env.test.2.0
# packaging==21.3
# pkgutil_resolve_name==.env.test.3.10
# pluggy==.env.test.0.0
# protobuf3-to-dict==0..env.test.5
# py==.env.test.11.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.21
# pyflakes==2.5.0
PyHamcrest==2.0.4
# pyparsing==3.0.9
# pypdf==3.2..env.test
# pyrsistent==0.19.2
# PySocks==.env.test.7..env.test
pytest==7.3
pytest-xdist==3.2
# python-dateutil==2.8.2
# python-dotenv==0.21.0
# requests==2.28..env.test
# rsa==4.9
# selene==2.0.0rc2
selenium==4.9.0
# six==.env.test.16.0
# smpplib==2.2..env.test
# sniffio==.env.test.3.0
# sortedcontainers==2.4.0
testit-adapter-pytest==2.0.10
# tomli==2.0..env.test
# tqdm==4.64..env.test
# trio==0.22.0
# trio-websocket==0.9.2
# typing_extensions==4.3.0
# urllib3==.env.test.26.13
# urllib3-secure-extra==0..env.test.0
# webdriver-manager==3.8.6
# wsproto==.env.test.2.0
zipp==3.11.0
curlify==2.2.1